Allow choice of map style
I only considered this for Apple/Google (not too worried about other map
apps) and in the end it's only possible for Apple maps.

Google maps will only let you choose satellite mode for directions but
then the map is zoomed out showing directions from your current
position which is not ideal, especially because I want to show satellite
view for the actual crag/sector positions (not the parking), for which
the navigation is unlikely to work well.